+++ src/components/sandbox/index.ts
@@ -1,6 +1,6 @@
 import Vue, { PluginObject } from 'vue';
 
-import RadioSandboxPlugin from '../radio/sandbox';
+import RadioSandboxPlugin from '../radio/radio-sandbox';
 
 const SandboxesPlugin: PluginObject<any> = {
     install(v, options): void {
